src/com/company/chapter02/ShopChapter02.java
package com.company.chapter02;

import java.util.ArrayList;
import java.util.List;
import java.util.function.Predicate;

import com.company.chapter01.Apple;

public class ShopChapter02 {
    public static List<Apple> filterApples(List<Apple> inventory, ApplePredicate p) {
        List<Apple> result = new ArrayList<>();

        for (Apple apple : inventory) {
            if (p.test(apple)) {
                result.add(apple);
            }
        }

        return result;
    }

    public static <T> List<T> filter(List<T> list, Predicate<T> p) {
        List<T> result = new ArrayList<>();

        for (T e : list) {
            if (p.test(e)) {
                result.add(e);
            }
        }

        return result;
    }
}
